refactor(cli): split main.rs into focused modules (RFC 0005, PR 8c)
Split the 1,159-line main.rs into idiomatic Rust modules:
- types.rs: API request/response types, shared validation helpers, colorize_status
- commands/workflow.rs: cmd_run, cmd_status
- commands/agent.rs: cmd_agent_list, cmd_agent_info, cmd_agent_reload, cmd_test, cmd_test_persona
- commands/logs.rs: cmd_logs
- commands/validate.rs: cmd_validate, find_validator_script, find_python_binary
- main.rs: clap definitions, main(), exhaustive command dispatch (188 lines)

All 20 tests pass, cargo clippy clean, cargo build --release succeeds.
No behavioral changes — pure file restructuring.
