+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-14  
-> **191** tracked files
+> **197** tracked files
 
 ## Summary
 
@@ -14,8 +14,8 @@
 | `(root)` | 20 |
 | `scripts/` | 12 |
 | `.github/` | 11 |
+| `cli/` | 9 |
 | `config/` | 9 |
-| `cli/` | 3 |
 | `schemas/` | 3 |
 | `blueprints/` | 2 |
 | `proto/` | 2 |
@@ -30,6 +30,7 @@
 | `.md` | 36 |
 | `.go` | 35 |
 | `.yaml` | 22 |
+| `.rs` | 7 |
 | `(no extension)` | 5 |
 | `.toml` | 4 |
 | `.json` | 4 |
@@ -40,7 +41,6 @@
 | `.orchestrator` | 1 |
 | `.pyi` | 1 |
 | `.lock` | 1 |
-| `.rs` | 1 |
 | `.mod` | 1 |
 | `.sum` | 1 |
 
@@ -127,7 +127,14 @@
 │   ├── Cargo.lock
 │   ├── Cargo.toml
 │   └── src/
-│       └── main.rs
+│       ├── commands/
+│       │   ├── agent.rs
+│       │   ├── logs.rs
+│       │   ├── mod.rs
+│       │   ├── validate.rs
+│       │   └── workflow.rs
+│       ├── main.rs
+│       └── types.rs
 ├── cliff.toml
 ├── cmd/
 │   └── orchestrator/
