+++ cli/src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod agent;
+pub(crate) mod logs;
+pub(crate) mod validate;
+pub(crate) mod workflow;
